Fix loop on aarch64

+++ src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp
@@ -5122,7 +5122,7 @@ MacroAssembler::KlassDecodeMode MacroAssembler::klass_decode_mode() {
   if (operand_valid_for_logical_immediate(
         /*is32*/false, (uint64_t)CompressedKlassPointers::base())) {
     const uint64_t range_mask =
-        nth_bit(ceil_log2(CompressedKlassPointers::highest_valid_narrow_klass_id()) +
+        nth_bit(ceil_log2((uint64_t)CompressedKlassPointers::highest_valid_narrow_klass_id()) +
                 CompressedKlassPointers::shift()) - 1;
     if (((uint64_t)CompressedKlassPointers::base() & range_mask) == 0) {
       return (_klass_decode_mode = KlassDecodeXor);
